+++ crates/services/gas_price_service/src/ports.rs
@@ -1,0 +1,14 @@
+use fuel_core_storage::Result as StorageResult;
+use fuel_core_types::{
+    blockchain::block::Block,
+    fuel_tx::Transaction,
+    fuel_types::BlockHeight,
+};
+
+pub trait L2Data: Send + Sync {
+    fn latest_height(&self) -> StorageResult<BlockHeight>;
+    fn get_block(
+        &self,
+        height: &BlockHeight,
+    ) -> StorageResult<Option<Block<Transaction>>>;
+}